pyproject.toml
[project]
name = "lang"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "discord-py>=2.5.2",
    "langchain-community>=0.3.27",
    "langchain-openai>=0.3.28",
    "langchain-qdrant>=0.2.0",
    "langchain>=0.3.26",
    "pydantic-settings>=2.10.1",
    "pydantic>=2.11.7",
    "qdrant-client>=1.15.0",
]

[dependency-groups]
dev = [
    "ruff>=0.12.4",
]
